LP2120581: Fix configure check for libdbi pgsql driver
When building with recent versions of gcc, the configure check for the
libdbi pgsql driver errors because of an implicit function definition
of the dbi_initialize function.  This is corrected by adding an
include of the dbi.h header to the prologue of the code generated for
the test.

Testing this is simple.  The configure step will fail with the following error on Debian 13:

checking for libdbi pgsql driver (dynamic load)... configure: error: in '/home/opensrf/Evergreen':
configure: error: "pgsql driver not installed?"
See 'config.log' for more details

After applying this patch and running autoreconf -i again, the
configure will continues past this and then fail on libpcre checks.
(See Launchpdad bug 2126962
https://bugs.launchpad.net/evergreen/+bug/2126962.)

The configure script should still work on all supported Linux distributions after applying the patch.

Release-note: Fixes an installation issue when checking for a
PostgreSQL driver.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>